Changed Test-Connection to Test-NetConnection
Test-Connection can fail as ICMP can be blocked.  Due to PowerShell 5.x limitations -TcpPort is only available PS7.x so we cannot specify a port as it may not work when using PS5.x which is the minimum version we can expect in production.

We can get around this by using Test-NetConnection which gives us access to -Port where we can use port 443 as port 80 is blocked in my testing.